fix: 첫 회원가입 사용자 로그인시 발생하던 오류 수정
